Update windowsservercore Docker image
Qt is updated to 5.12.5, MSVC is updated to 2017. Qbs is updated to
1.15.0.
Docker-compose is used for managing the image. A unique tag name is used
instead
of the generic windowsservercore tag to be always able to refer to a
specific
version. Docker-compose is now used for managing the image.

Open items:
- Make clcache work properly
- Fix doc build
- Fix joblimit test
- Decide about zip/unzip
- Upload image

Change-Id: Ia5733992e7692561b1a6e74a4ac2ea3a7a01c4d8